Update dispatch clean with flexible definitions
