--- rmf_demos_tasks/rmf_demos_tasks/dispatch_multi_clean.py
@@ -1,127 +1,0 @@
-#!/usr/bin/env python3
-
-# Copyright 2022 Open Source Robotics Foundation, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import sys
-import uuid
-import argparse
-import json
-
-import rclpy
-from rclpy.node import Node
-from rclpy.parameter import Parameter
-from rclpy.qos import qos_profile_system_default
-from rclpy.qos import QoSProfile
-from rclpy.qos import QoSHistoryPolicy as History
-from rclpy.qos import QoSDurabilityPolicy as Durability
-from rclpy.qos import QoSReliabilityPolicy as Reliability
-
-from rmf_task_msgs.msg import ApiRequest
-
-
-###############################################################################
-
-class TaskRequester(Node):
-
-    def __init__(self, argv=sys.argv):
-        super().__init__('task_requester')
-        parser = argparse.ArgumentParser()
-        parser.add_argument('-z', '--zones', required=True,
-                            type=str, nargs='+', help='List of cleaning zones')
-        parser.add_argument('-F', '--fleet', required=False, default='',
-                            type=str, help='Fleet name')
-        parser.add_argument('-R', '--robot', required=False, default='',
-                            type=str, help='Robot name')
-        parser.add_argument('-st', '--start_time',
-                            help='Start time from now in secs, default: 0',
-                            type=int, default=0)
-        parser.add_argument('-pt', '--priority',
-                            help='Priority value for this request',
-                            type=int, default=0)
-        parser.add_argument("--use_sim_time", action="store_true",
-                            help='Use sim time, default: false')
-
-        self.args = parser.parse_args(argv[1:])
-
-        transient_qos = QoSProfile(
-            history=History.KEEP_LAST,
-            depth=1,
-            reliability=Reliability.RELIABLE,
-            durability=Durability.TRANSIENT_LOCAL)
-
-        self.pub = self.create_publisher(
-          ApiRequest, 'task_api_requests', transient_qos)
-
-        # enable ros sim time
-        if self.args.use_sim_time:
-            self.get_logger().info("Using Sim Time")
-            param = Parameter("use_sim_time", Parameter.Type.BOOL, True)
-            self.set_parameters([param])
-
-        # Construct task
-        msg = ApiRequest()
-        msg.request_id = "multiclean_" + str(uuid.uuid4())
-        payload = {}
-        if self.args.fleet and self.args.robot:
-            payload["type"] = "robot_task_request"
-            payload["robot"] = self.args.robot
-            payload["fleet"] = self.args.fleet
-        else:
-            payload["type"] = "dispatch_task_request"
-        request = {}
-
-        # Set task request start time
-        now = self.get_clock().now().to_msg()
-        now.sec = now.sec + self.args.start_time
-        start_time = now.sec * 1000 + round(now.nanosec/10**6)
-        request["unix_millis_earliest_start_time"] = start_time
-
-        # Define task request category
-        request["category"] = "compose"
-
-        # Define task request description with phases
-        description = {}  # task_description_Compose.json
-        description["category"] = "multi_clean"
-        description["phases"] = []
-        activities = []
-        # Add each cleaning activity
-        for zone in self.args.zones:
-            activities.append({"category": "clean",
-                               "description": {"zone": zone}})
-        # Add activities to phases
-        description["phases"].append(
-            {"activity": {"category": "sequence",
-                          "description": {"activities": activities}}})
-
-        request["description"] = description
-        payload["request"] = request
-        msg.json_msg = json.dumps(payload)
-
-        self.pub.publish(msg)
-
-
-###############################################################################
-
-
-def main(argv=sys.argv):
-    rclpy.init(args=sys.argv)
-    args_without_ros = rclpy.utilities.remove_ros_args(sys.argv)
-
-    task_requester = TaskRequester(args_without_ros)
-    rclpy.shutdown()
-
-
-if __name__ == '__main__':
-    main(sys.argv)
